local/spectrumscale/spectrumscale.go
/**
 * Copyright 2017 IBM Corp.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package spectrumscale

import (
    "github.com/IBM/ubiquity/utils/logs"
	"github.com/IBM/ubiquity/utils"
	"fmt"
	"github.com/IBM/ubiquity/local/spectrumscale/connectors"
	"sync"
	"github.com/IBM/ubiquity/resources"
)

type spectrumLocalClient struct {
	logger         logs.Logger
	connector      connectors.SpectrumScaleConnector
	dataModel      SpectrumDataModelWrapper
	executor       utils.Executor
	isActivated    bool
	isMounted      bool
	config         resources.SpectrumScaleConfig
	activationLock *sync.RWMutex
}

const (
	Type            string = "type"
	TypeFileset     string = "fileset"

	FilesetID string = "fileset"
	Directory string = "directory"
	Quota     string = "quota"

	Filesystem string = "filesystem"

	IsPreexisting string = "isPreexisting"

    SpectrumScaleConfigUser = "REST_USER"
    SpectrumScaleConfigPassword = "REST_PASSWORD"
    SpectrumScaleConfigFilesystem = "DEFAULT_FILESYSTEM_NAME"
)

func NewSpectrumLocalClient(config resources.UbiquityServerConfig) (resources.StorageClient, error) {
	if config.SpectrumScaleConfig.DefaultFilesystemName == "" {
		return nil, fmt.Errorf("spectrumLocalClient: init: missing required parameter 'spectrumDefaultFileSystem'")
	}

	return newSpectrumLocalClient(config.SpectrumScaleConfig, resources.SpectrumScale)
}

func NewSpectrumLocalClientWithConnectors(logger logs.Logger, connector connectors.SpectrumScaleConnector, spectrumExecutor utils.Executor, config resources.SpectrumScaleConfig, datamodel SpectrumDataModelWrapper) (resources.StorageClient, error) {
	return &spectrumLocalClient{logger: logger, connector: connector, dataModel: datamodel, executor: spectrumExecutor, config: config, activationLock: &sync.RWMutex{}}, nil
}

func newSpectrumLocalClient(config resources.SpectrumScaleConfig, backend string) (*spectrumLocalClient, error) {
    logger := logs.GetLogger()
    defer logger.Trace(logs.DEBUG)()

    // Validate required parameters
    if err := validateSpectrumscaleConfig(logger, config); err != nil {
        return &spectrumLocalClient{}, err
    }

	client, err := connectors.GetSpectrumScaleConnector(logger, config)
	if err != nil {
		return &spectrumLocalClient{}, logger.ErrorRet(err, "GetSpectrumScaleConnector failed")
	}

	datamodel := NewSpectrumDataModelWrapper(backend)

	dbName := datamodel.GetDbName()

	SpectrumScaleLocalClient := &spectrumLocalClient{logger: logger, connector: client, dataModel: datamodel, config: config, executor: utils.NewExecutor(), activationLock: &sync.RWMutex{}}


    // Validate Spectrum Scale cluster
    if err := basicSpectrumscaleLocalClientValidation(SpectrumScaleLocalClient); err != nil {
        return &spectrumLocalClient{}, err
    }

    volume, err := client.ListFileset(config.DefaultFilesystemName, dbName)
	if err == nil {
		logger.Debug("DB volume fileset present")
        scaleDbVol := &SpectrumScaleVolume{Volume: resources.Volume{Name: volume.Name, Backend: backend}, Type: Fileset, FileSystem: config.DefaultFilesystemName, Fileset: dbName}
        datamodel.UpdateDatabaseVolume(scaleDbVol)
	} else {
		logger.Debug("DB Vol Fileset Not Found", logs.Args{{"Filesystem", config.DefaultFilesystemName}, {"Fileset", dbName}})

	}
	return SpectrumScaleLocalClient, nil
}

func validateSpectrumscaleConfig(logger logs.Logger, config resources.SpectrumScaleConfig) error {
    defer logger.Trace(logs.DEBUG)()

    if config.RestConfig.User == ""{
        return logger.ErrorRet(&SpectrumScaleConfigError{ConfigParam: resources.SpectrumScaleParamPrefix + SpectrumScaleConfigUser }, "")
    }

    if config.RestConfig.Password == ""{
        return logger.ErrorRet(&SpectrumScaleConfigError{ConfigParam: resources.SpectrumScaleParamPrefix + SpectrumScaleConfigPassword}, "")
    }

    if config.DefaultFilesystemName == ""{
        return logger.ErrorRet(&SpectrumScaleConfigError{ConfigParam: resources.SpectrumScaleParamPrefix + SpectrumScaleConfigFilesystem}, "")
    }
    return nil
}

func basicSpectrumscaleLocalClientValidation(spectrumScaleClient *spectrumLocalClient) error {
    defer spectrumScaleClient.logger.Trace(logs.DEBUG)()

    // SpectrumScale Validation
    // 1. get cluster id with provided configuration to confirm connectivity and credentials
    spectrumScaleClient.logger.Info("Verifying the Connectivity and Credential of Spectrum Scale Cluster by getting Cluster ID of Spectrum Scale Cluster")
    clusterid, err := spectrumScaleClient.connector.GetClusterId()
    if err != nil {
        return spectrumScaleClient.logger.ErrorRet(&SpectrumScaleGetClusterIdError{ErrorMsg: SpectrumScaleGetClusterIdErrorStr}, "")
    }
    spectrumScaleClient.logger.Info("Cluster ID of SpectrumScale Cluster", logs.Args{{"Cluster ID", clusterid}})

    // 2. Check if default filesystem exist and mounted
    spectrumScaleClient.logger.Info("Check if filesystem exist and mounted", logs.Args{{"Filesystem", spectrumScaleClient.config.DefaultFilesystemName}})
    isfsmounted, err := spectrumScaleClient.connector.IsFilesystemMounted(spectrumScaleClient.config.DefaultFilesystemName)
    if err != nil {
        return spectrumScaleClient.logger.ErrorRet(&SpectrumScaleFileSystemNotPresent{Filesystem: spectrumScaleClient.config.DefaultFilesystemName},"")
    }
    if ! isfsmounted {
        return spectrumScaleClient.logger.ErrorRet(&SpectrumScaleFileSystemNotMounted{Filesystem: spectrumScaleClient.config.DefaultFilesystemName},"")
    }
    spectrumScaleClient.logger.Info("Spectrum Scale filesystem is mounted.", logs.Args{{"Filesystem", spectrumScaleClient.config.DefaultFilesystemName}})

    return nil
}

func (s *spectrumLocalClient) Attach(attachRequest resources.AttachRequest) (volumeMountpoint string, err error) {
	return "", nil
}

func (s *spectrumLocalClient) Detach(detachRequest resources.DetachRequest) (err error) {
	return nil
}

func (s *spectrumLocalClient) Activate(activateRequest resources.ActivateRequest) (err error) {
    defer s.logger.Trace(logs.DEBUG)()

	s.activationLock.RLock()
	if s.isActivated {
		s.activationLock.RUnlock()
		return nil
	}
	s.activationLock.RUnlock()

	s.activationLock.Lock() //get a write lock to prevent others from repeating these actions
	defer s.activationLock.Unlock()

	s.isActivated = true
	return nil
}

func (s *spectrumLocalClient) CreateVolume(createVolumeRequest resources.CreateVolumeRequest) (err error) {
    defer s.logger.Trace(logs.DEBUG)()

	_, volExists, err := s.dataModel.GetVolume(createVolumeRequest.Name)

	if err != nil {
		return s.logger.ErrorRet(err, "Failed to get volume details from Database", logs.Args{{"name", createVolumeRequest.Name}})
	}

	if volExists {
		return s.logger.ErrorRet(fmt.Errorf("Volume already exists"),"")
	}

	s.logger.Debug("Opts for create:", logs.Args{{"Opts", createVolumeRequest.Opts}})

	if len(createVolumeRequest.Opts) == 0 {
        return s.createFilesetVolume(s.config.DefaultFilesystemName, createVolumeRequest.Name, createVolumeRequest.Opts)
	}
	s.logger.Debug("Trying to determine type for request")
	userSpecifiedType, err := determineTypeFromRequest(s.logger, createVolumeRequest.Opts)
	if err != nil {
		return s.logger.ErrorRet(err, "Error determining type")
	}
	s.logger.Debug("Volume type requested", logs.Args{{"userSpecifiedType",userSpecifiedType}})
	isExistingVolume, filesystem, existingFileset, err := s.validateAndParseParams(s.logger, createVolumeRequest.Opts)
	if err != nil {
		return s.logger.ErrorRet(err, "Error in validate params")
	}


	if isExistingVolume && userSpecifiedType == TypeFileset {
		quota, quotaSpecified := createVolumeRequest.Opts[Quota]
		if quotaSpecified {
			return s.updateDBWithExistingFilesetQuota(filesystem, createVolumeRequest.Name, existingFileset, quota.(string), createVolumeRequest.Opts)
		}
		return s.updateDBWithExistingFileset(filesystem, createVolumeRequest.Name, existingFileset, createVolumeRequest.Opts)
	}

	if userSpecifiedType == TypeFileset {
		quota, quotaSpecified := createVolumeRequest.Opts[Quota]
        if quotaSpecified {
            return s.createFilesetQuotaVolume(filesystem, createVolumeRequest.Name, quota.(string), createVolumeRequest.Opts)
        }
		return s.createFilesetVolume(filesystem, createVolumeRequest.Name, createVolumeRequest.Opts)
	}
	return s.logger.ErrorRet(fmt.Errorf("Internal error"),"")
}

func (s *spectrumLocalClient) RemoveVolume(removeVolumeRequest resources.RemoveVolumeRequest) (err error) {
    defer s.logger.Trace(logs.DEBUG)()

	existingVolume, volExists, err := s.dataModel.GetVolume(removeVolumeRequest.Name)

	if err != nil {
		return s.logger.ErrorRet(err, "Unable to get volume from Database", logs.Args{{"VolumeName", removeVolumeRequest.Name}})
	}

	if volExists == false {
		return &resources.VolumeNotFoundError{VolName: removeVolumeRequest.Name}
	}

	isFilesetLinked, err := s.connector.IsFilesetLinked(existingVolume.FileSystem, existingVolume.Fileset)

	if err != nil {
		return s.logger.ErrorRet(err, "Unable to check if fileset is linked", logs.Args{{"Filesystem", existingVolume.FileSystem}, {"Fileset", existingVolume.Fileset}})
	}

    if isFilesetLinked && existingVolume.IsPreexisting == false {
        err := s.connector.UnlinkFileset(existingVolume.FileSystem, existingVolume.Fileset)
		if err != nil {
			return s.logger.ErrorRet(err, "Failed to Unlink fileset", logs.Args{{"Filesystem", existingVolume.FileSystem}, {"Fileset", existingVolume.Fileset}})
		}
	}
	err = s.dataModel.DeleteVolume(removeVolumeRequest.Name)

	if err != nil {
		return s.logger.ErrorRet(err, "failed to delete volume", logs.Args{{"VolumeName", removeVolumeRequest.Name}})
	}
	if s.config.ForceDelete == true && existingVolume.IsPreexisting == false {
		err = s.connector.DeleteFileset(existingVolume.FileSystem, existingVolume.Fileset)

		if err != nil {
			return s.logger.ErrorRet(err, "failed to delete fileset", logs.Args{{"Filesystem", existingVolume.FileSystem}, {"Fileset", existingVolume.Fileset}})
		}
	}

	return nil
}

func (s *spectrumLocalClient) GetVolume(getVolumeRequest resources.GetVolumeRequest) (resources.Volume, error) {
    defer s.logger.Trace(logs.DEBUG)()

	existingVolume, volExists, err := s.dataModel.GetVolume(getVolumeRequest.Name)
	if err != nil {
		return resources.Volume{}, err
	}
	if volExists == false {
		return resources.Volume{},&resources.VolumeNotFoundError{VolName: getVolumeRequest.Name}
	}

	return resources.Volume{Name: existingVolume.Volume.Name, Backend: existingVolume.Volume.Backend, Mountpoint: existingVolume.Volume.Mountpoint}, nil
}

func (s *spectrumLocalClient) GetVolumeConfig(getVolumeConfigRequest resources.GetVolumeConfigRequest) (volumeConfigDetails map[string]interface{}, err error) {
    defer s.logger.Trace(logs.DEBUG)()

	existingVolume, volExists, err := s.dataModel.GetVolume(getVolumeConfigRequest.Name)

	if err != nil {
		return nil, s.logger.ErrorRet(err, "GetVolume failed", logs.Args{{"VolumeName", getVolumeConfigRequest.Name}})
	}

	if volExists {
		volumeConfigDetails = make(map[string]interface{})
		volumeMountpoint, err := s.getVolumeMountPoint(existingVolume)
		if err != nil {
			return nil, s.logger.ErrorRet(err, "failed to get mountpoint for volume", logs.Args{{"VolumeName", getVolumeConfigRequest.Name}})
		}

		isFilesetLinked, err := s.connector.IsFilesetLinked(existingVolume.FileSystem, existingVolume.Fileset)
		if err != nil {
			return nil, s.logger.ErrorRet(err, "failed to check if fileset is linked", logs.Args{{"Filesystem", existingVolume.FileSystem}, {"Fileset", existingVolume.Fileset}})
		}
		if isFilesetLinked {
			volumeConfigDetails["mountpoint"] = volumeMountpoint
		}

		volumeConfigDetails[FilesetID] = existingVolume.Fileset
		volumeConfigDetails[Filesystem] = existingVolume.FileSystem
		if existingVolume.GID != "" {
			volumeConfigDetails[UserSpecifiedGID] = existingVolume.GID
		}
		if existingVolume.UID != "" {
			volumeConfigDetails[UserSpecifiedUID] = existingVolume.UID
		}
		volumeConfigDetails[IsPreexisting] = existingVolume.IsPreexisting
		volumeConfigDetails[Type] = existingVolume.Type

		return volumeConfigDetails, nil
	}
	return nil, &resources.VolumeNotFoundError{VolName: getVolumeConfigRequest.Name}
}

func (s *spectrumLocalClient) ListVolumes(listVolumesRequest resources.ListVolumesRequest) ([]resources.Volume, error) {
    defer s.logger.Trace(logs.DEBUG)()

	var err error
	volumesInDb, err := s.dataModel.ListVolumes()
	if err != nil {
		return nil, s.logger.ErrorRet(err, "failed to list volumes")
	}
	return volumesInDb, nil
}

func (s *spectrumLocalClient) createFilesetVolume(filesystem, name string, opts map[string]interface{}) error {
    defer s.logger.Trace(logs.DEBUG)()

	filesetName := generateFilesetName(name)

	var dbVolerr error
	var err error

    err = s.checkIfFSMounted(filesystem)
    if err != nil {
        return err
    }

	isDbVolume := s.dataModel.IsDbVolume(name)
	if isDbVolume {
        // Check if fileset is present
        s.logger.Debug("DB Volume, check if present",logs.Args{{"VolumeName",name},{"Filesystem",filesystem}})
		_, dbVolerr = s.connector.ListFileset(filesystem, name)
	}

	if !isDbVolume || dbVolerr != nil {
        s.logger.Debug("Create fileset: DB volume is not present OR it is normal volume creation", logs.Args{{"VolumeName", filesetName}, {"Filesystem", filesystem}})
		err := s.connector.CreateFileset(filesystem, filesetName, opts)
		if  err != nil {
			return s.logger.ErrorRet(err, "Error creating fileset", logs.Args{{"Filesystem", filesystem}, {"Fileset", filesetName}})
		}
	}

	err = s.dataModel.InsertFilesetVolume(filesetName, name, filesystem, false, opts)
	if err != nil {
		return s.logger.ErrorRet(err, "Error inserting fileset", logs.Args{{"Filesystem", filesystem}, {"Fileset", filesetName}})
	}

	s.logger.Debug("Created fileset volume with fileset ", logs.Args{{"filesetName", filesetName}})
	return nil
}

func (s *spectrumLocalClient) createFilesetQuotaVolume(filesystem, name, quota string, opts map[string]interface{}) error {
    defer s.logger.Trace(logs.DEBUG)()

	filesetName := generateFilesetName(name)
    var dbVolerr error
	var err error

    err = s.checkIfFSMounted(filesystem)
    if err != nil {
        return err
    }

    err = s.connector.CheckIfFSQuotaEnabled(filesystem)
    if err != nil {
        return s.logger.ErrorRet(&SpectrumScaleQuotaNotEnabledError{Filesystem: filesystem}, "")
    }
    if s.dataModel.IsDbVolume(name) {
    //Check if fileset is present
    s.logger.Debug("DB Volume, check if present",logs.Args{{"VolumeName",name},{"Filesystem",filesystem}})
        _, dbVolerr = s.connector.ListFileset(filesystem, name)
    }

    if !s.dataModel.IsDbVolume(name) || dbVolerr != nil {
        s.logger.Debug("Create fileset: DB volume is not present OR it is normal volume creation", logs.Args{{"VolumeName", filesetName}, {"Filesystem", filesystem}})
        err := s.connector.CreateFileset(filesystem, filesetName, opts)
        if  err != nil {
            return s.logger.ErrorRet(err, "Error creating fileset", logs.Args{{"Filesystem", filesystem}, {"Fileset", filesetName}})
        }

        err = s.connector.SetFilesetQuota(filesystem, filesetName, quota)
		if err != nil {
			deleteErr := s.connector.DeleteFileset(filesystem, filesetName)
			if deleteErr != nil {
				return s.logger.ErrorRet(deleteErr, "Error setting quota (rollback error on delete fileset", logs.Args{{"filesetName", filesetName}})
			}
			return err
		}
	}

	err = s.dataModel.InsertFilesetQuotaVolume(filesetName, quota, name, filesystem, false, opts)

	if err != nil {
		return err
	}

	s.logger.Debug("Created fileset volume with fileset quota ", logs.Args{{"filesetName", filesetName},{"quota", quota}})
	return nil
}

func generateFilesetName(name string) string {
	//TODO: placeholder for now
	return name
	//return strconv.FormatInt(time.Now().UnixNano(), 10)
}

//TODO move updates to DB file

func (s *spectrumLocalClient) updateDBWithExistingFileset(filesystem, name, userSpecifiedFileset string, opts map[string]interface{}) error {
    defer s.logger.Trace(logs.DEBUG)()

	s.logger.Debug("User specified fileset", logs.Args{{"userSpecifiedFileset", userSpecifiedFileset}})

	_, err := s.connector.ListFileset(filesystem, userSpecifiedFileset)
	if err != nil {
		return s.logger.ErrorRet(err, "Fileset does not exist", logs.Args{{"filesystem", filesystem}, {"Fileset", userSpecifiedFileset}})
	}

	err = s.dataModel.InsertFilesetVolume(userSpecifiedFileset, name, filesystem, true, opts)

	if err != nil {
		return s.logger.ErrorRet(err, "InsertFilesetVolume failed", logs.Args{{"filesystem", filesystem}, {"Fileset", userSpecifiedFileset}})
	}
	return nil
}

func (s *spectrumLocalClient) checkIfVolumeExistsInDB(name, userSpecifiedFileset string) error {
    defer s.logger.Trace(logs.DEBUG)()
	getVolumeConfigRequest := resources.GetVolumeConfigRequest{Name: name}
	volumeConfigDetails, err := s.GetVolumeConfig(getVolumeConfigRequest)

	if err != nil {
		return s.logger.ErrorRet(err, "GetVolumeConfig failed", logs.Args{{"VolumeName", name}})
	}

	if volumeConfigDetails["FilesetId"] != userSpecifiedFileset {
		return s.logger.ErrorRet(fmt.Errorf("volume %s with fileset %s not found", name, userSpecifiedFileset),"")
	}
	return nil
}

func (s *spectrumLocalClient) updateDBWithExistingFilesetQuota(filesystem, name, userSpecifiedFileset, quota string, opts map[string]interface{}) error {
    defer s.logger.Trace(logs.DEBUG)()

	filesetQuota, err := s.connector.ListFilesetQuota(filesystem, userSpecifiedFileset)

	if err != nil {
		return s.logger.ErrorRet(err, "ListFilesetQuota failed",logs.Args{{"filesystem", filesystem}, {"Fileset", userSpecifiedFileset}})
	}

	s.logger.Debug("For REST connector converting quotas to bytes")
    filesetQuotaBytes, err := utils.ConvertToBytes(s.logger, filesetQuota)
	if err != nil {
		return s.logger.ErrorRet(err, "utils.ConvertToBytes failed", logs.Args{{"filesetQuota", filesetQuota}})
	}

	quotasBytes, err := utils.ConvertToBytes(s.logger, quota)
	if err != nil {
        s.logger.Debug("utils.ConvertToBytes failed ", logs.Args{{"Error", err}})
		return s.logger.ErrorRet(err, "utils.ConvertToBytes failed ", logs.Args{{"Fileset Quota", quota}})
	}

    if filesetQuotaBytes < quotasBytes {
        return s.logger.ErrorRet(fmt.Errorf("user-specified quota %v is greater than listed quota %v for fileset %s", quotasBytes, filesetQuotaBytes, userSpecifiedFileset),"")
	}

	err = s.dataModel.InsertFilesetQuotaVolume(userSpecifiedFileset, quota, name, filesystem, true, opts)

	if err != nil {
		s.logger.Debug("", logs.Args{{"Error", err.Error()}})
		return s.logger.ErrorRet(err, "InsertFilesetQuotaVolume failed")
	}
    return nil
}

func determineTypeFromRequest(logger logs.Logger, opts map[string]interface{}) (string, error) {
    defer logger.Trace(logs.DEBUG)()

	userSpecifiedType, exists := opts[Type]
	if exists == false {
		return TypeFileset, nil
	}

	if userSpecifiedType.(string) != TypeFileset {
		return "", logger.ErrorRet(fmt.Errorf("Unknown 'type' = %s specified", userSpecifiedType.(string)), "")
	}

	return userSpecifiedType.(string), nil
}

func (s *spectrumLocalClient) checkIfFSMounted(filesystem string) error {
    defer s.logger.Trace(logs.DEBUG)()

    isfsmounted, err := s.connector.IsFilesystemMounted(filesystem)
    if err != nil {
        return s.logger.ErrorRet(&SpectrumScaleFileSystemMountError{Filesystem: filesystem},"")
    }

    if !isfsmounted {
        return s.logger.ErrorRet(&SpectrumScaleFileSystemNotMounted{Filesystem: filesystem},"")
    }
    return nil
}

func (s *spectrumLocalClient) validateAndParseParams(logger logs.Logger, opts map[string]interface{}) (bool, string, string, error) {
    defer s.logger.Trace(logs.DEBUG)()

	existingFileset, existingFilesetSpecified := opts[TypeFileset]
	filesystem, filesystemSpecified := opts[Filesystem]
	_, uidSpecified := opts[UserSpecifiedUID]
	_, gidSpecified := opts[UserSpecifiedGID]

	userSpecifiedType, err := determineTypeFromRequest(logger, opts)
	if err != nil {
		return false, "", "", s.logger.ErrorRet(err, "failed to determine type")
	}

	if uidSpecified && gidSpecified {
		if existingFilesetSpecified{
			return true, "", "", s.logger.ErrorRet(fmt.Errorf("uid/gid cannot be specified along with existing fileset"), "")
		}
	}

	if (userSpecifiedType == TypeFileset && existingFilesetSpecified){
		if filesystemSpecified == false {
			logger.Debug("'filesystem' is a required opt for using existing volumes")
			return true, filesystem.(string), existingFileset.(string), s.logger.ErrorRet(fmt.Errorf("'filesystem' is a required opt for using existing volumes"), "")
		}
		logger.Debug("Valid: existing FILESET")
		return true, filesystem.(string), existingFileset.(string), nil

	} else if filesystemSpecified == false {
		return false, s.config.DefaultFilesystemName, "", nil

		}else {
			return false, filesystem.(string), "", nil
		}
}

func (s *spectrumLocalClient) getVolumeMountPoint(volume SpectrumScaleVolume) (string, error) {
    defer s.logger.Trace(logs.DEBUG)()

    vol, err := s.connector.ListFileset(volume.FileSystem, volume.Fileset)
	if err != nil {
		return "", err
	}
    s.logger.Debug("Volume mount point:",logs.Args{{"volume filesystem:", volume.FileSystem},{"volume fileset:", volume.Fileset}, {"volume mountpoint:",vol.Mountpoint}})
    return vol.Mountpoint, nil
}
